requirements.txt
chromadb==0.5.23
sentence-transformers==3.3.1
anthropic==0.40.0
python-dotenv==1.0.1
slack-bolt==1.21.3
